Layouts and navigation
Common
UI layer
-extend package structure
-rename package login to drivelogin
-add ViewModel classes

Android
UI layer
-move HerdrActivity to ui.main
-add Fragment classes in their respective packages
