+++ app/build.gradle
@@ -81,7 +81,7 @@ kotlin {
                 implementation "org.jetbrains.kotlinx:kotlinx-coroutines-core-common:$coroutineVersion"
 
                 // MOKO - MVVM
-               implementation "dev.icerock.moko:mvvm:$mokoMvvmVersion"
+                implementation "dev.icerock.moko:mvvm:$mokoMvvmVersion"
 
                 // KODEIN
                 implementation "org.kodein.di:kodein-di-core:$kodeinVersion"
@@ -96,8 +96,8 @@ kotlin {
         }
         commonTest {
             dependencies {
-        		implementation kotlin('test-common')
-        		implementation kotlin('test-annotations-common')
+                implementation kotlin('test-common')
+                implementation kotlin('test-annotations-common')
             }
         }
         androidMain {
@@ -106,6 +106,7 @@ kotlin {
                 implementation "org.jetbrains.kotlin:kotlin-stdlib-common:1.3.70"
                 implementation fileTree(dir: 'libs', include: ['*.jar'])
                 implementation "androidx.appcompat:appcompat:$appCompatVersion"
+                implementation "androidx.fragment:fragment:$fragmentVersion"
                 implementation "androidx.constraintlayout:constraintlayout:$constraintLayoutVersion"
 
                 //secure local storage
