+++ app/src/commonMain/kotlin/com/ludoscity/herdr/common/ui/drivesetup/DriveSetupViewModel.kt
@@ -16,18 +16,9 @@
  *     along with this program.  If not, see <https://www.gnu.org/licenses/>.
  */
 
-package com.ludoscity.herdr.ui
+package com.ludoscity.herdr.common.ui.drivesetup
 
-import androidx.lifecycle.ViewModel
-import androidx.lifecycle.ViewModelProvider
-import com.ludoscity.herdr.common.data.SecureDataStore
-import com.ludoscity.herdr.common.ui.login.LoginViewModel
+import dev.icerock.moko.mvvm.viewmodel.ViewModel
 
-class HerdrActivityModelFactory(private val dataStore: SecureDataStore) : ViewModelProvider.NewInstanceFactory() {
-
-    override fun <T : ViewModel> create(modelClass: Class<T>): T {
-
-        @Suppress("UNCHECKED_CAST")
-        return LoginViewModel(dataStore) as T
-    }
+class DriveSetupViewModel : ViewModel() {
 }
