+++ app/src/main/AndroidManifest.xml
@@ -28,7 +28,7 @@
             android:supportsRtl="true"
             android:theme="@style/AppTheme">
         <activity
-                android:name="com.ludoscity.herdr.ui.HerdrActivity"
+            android:name="com.ludoscity.herdr.ui.main.HerdrActivity"
                 android:launchMode="singleTop">
             <intent-filter>
                 <action android:name="android.intent.action.VIEW" />
