+++ build.gradle
@@ -42,11 +42,11 @@ ext {
     appCompatVersion = "1.1.0"
     constraintLayoutVersion = "1.1.3"
     coroutineVersion = "1.3.5"
+    fragmentVersion = "1.2.4"
     kodeinVersion = "6.5.3"
     ktorVersion = "1.3.2"
     mokoMvvmVersion = "0.6.0"
     serializerVersion = "0.20.0"
-    nimbusOauthSdkVersion = "6.13"
     appAuthVersion = "0.7.1"
     androidXBrowserVersion = "1.2.0"
     materialDesignVersion = "1.1.0"
